added read/write locks and file locks

+++ src/main/java/com/thorinhood/processors/actions/DeleteBucketProcessor.java
@@ -10,7 +10,6 @@ import io.netty.handler.codec.http.HttpResponseStatus;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import java.io.File;
 import java.util.Map;
 
 public class DeleteBucketProcessor extends Processor {
@@ -24,7 +23,7 @@ public class DeleteBucketProcessor extends Processor {
     @Override
     protected void processInner(ChannelHandlerContext context, FullHttpRequest request, ParsedRequest parsedRequest,
                                 Object... arguments) throws Exception {
-        checkRequest(parsedRequest, true);
+        checkRequestPermissions(parsedRequest, true);
         S3_DRIVER.deleteBucket(parsedRequest.getS3BucketPath());
         sendResponseWithoutContent(context, HttpResponseStatus.OK, request, Map.of(
                 "Date", DateTimeUtil.currentDateTime()
